Started unit testing for head.go, freelist.go, store.go

+++ btree/cow.go
@@ -31,7 +31,7 @@ func (kn *knode) split() (Node, Interface) {
 
     copy(newkn.ks[0:], kn.ks[max/2+1:])
     kn.ks = kn.ks[0:max/2+1]
-    kn.size = uint64( len(kn.ks) )
+    kn.size = len(kn.ks)
 
     copy(newkn.vs[0:], kn.vs[max/2+1:])
     kn.vs = append( kn.vs[0:max/2+1], 0 )
@@ -50,7 +50,7 @@ func (in *inode) split() (Node, Interface) {
     copy(newin.ks[0:], in.ks[max/2+1:])
     median := in.ks[max/2]
     in.ks = in.ks[0:max/2]
-    in.size = uint64( len(in.ks) )
+    in.size = len(in.ks)
 
     copy(newin.vs[0:], in.vs[max/2+1:])
     in.vs = in.vs[0:max/2+1]
@@ -75,10 +75,10 @@ func (kn *knode) merge(other_ Node, median bkey) Node {
     return kn
 }
 
-func (kn *knode) rotateRight(child_ Node, count uint64, median bkey) bkey {
+func (kn *knode) rotateRight(child_ Node, count int, median bkey) bkey {
     child := child_.(*knode)
-    chlen := uint64(len(child.ks))
-    knlen := uint64(len(kn.ks))
+    chlen := len(child.ks)
+    knlen := len(kn.ks)
     // Move last `count` keys from kn -> child.
     child.ks = child.ks[:chlen+count]   // First expand
     copy( child.ks[count:], child.ks[0:chlen] )
@@ -86,7 +86,7 @@ func (kn *knode) rotateRight(child_ Node, count uint64, median bkey) bkey {
     // Blindly shrink kn keys
     kn.ks = kn.ks[:knlen-count]
     // Update size.
-    kn.size, child.size = uint64(len(kn.ks)), uint64(len(child.ks))
+    kn.size, child.size = len(kn.ks), len(child.ks)
     // Move last count values from kn -> child
     child.vs = child.vs[:chlen+count+1] // First expand
     copy( child.vs[count:], child.vs[0:chlen+1] )
@@ -97,16 +97,16 @@ func (kn *knode) rotateRight(child_ Node, count uint64, median bkey) bkey {
     return child.ks[0]
 }
 
-func (child *knode) rotateLeft(kn_ Node, count uint64, median bkey) bkey {
+func (child *knode) rotateLeft(kn_ Node, count int, median bkey) bkey {
     kn := kn_.(*knode)
-    chlen := uint64(len(child.ks))
+    chlen := len(child.ks)
     // Move first `count` keys from kn -> child.
     child.ks = child.ks[:chlen+count]  // First expand
     copy( child.ks[chlen:], kn.ks[0:count] )
     // Blindly shrink kn keys
     kn.ks = kn.ks[count:]
     // Update size.
-    kn.size, child.size = uint64(len(kn.ks)), uint64(len(child.ks))
+    kn.size, child.size = len(kn.ks), len(child.ks)
     // Move last count values from kn -> child
     child.vs = child.vs[:chlen+count]    // First expand
     copy( child.vs[chlen:], kn.vs[0:count] )
@@ -133,11 +133,11 @@ func (in *inode) merge(other_ Node, median bkey) Node {
     return in
 }
 
-func (in *inode) rotateRight(child_ Node, count uint64, median bkey) bkey {
+func (in *inode) rotateRight(child_ Node, count int, median bkey) bkey {
     child := child_.(*inode)
     in.ks = append(in.ks, median)
-    chlen := uint64(len(child.ks))
-    inlen := uint64(len(in.ks))
+    chlen := len(child.ks)
+    inlen := len(in.ks)
     // Move last `count` keys from in -> child.
     child.ks = child.ks[:chlen+count]   // First expand
     copy( child.ks[count:], child.ks[0:chlen] )
@@ -145,7 +145,7 @@ func (in *inode) rotateRight(child_ Node, count uint64, median bkey) bkey {
     // Blindly shrink in keys
     in.ks = in.ks[:inlen-count]
     // Update size.
-    in.size, child.size = uint64(len(in.ks)), uint64(len(child.ks))
+    in.size, child.size = len(in.ks), len(child.ks)
     // Move last count values from in -> child
     child.vs = child.vs[:chlen+count+1] // First expand
     copy( child.vs[count:], child.vs[0:chlen+1] )
@@ -153,22 +153,22 @@ func (in *inode) rotateRight(child_ Node, count uint64, median bkey) bkey {
     // Pop out median
     median = in.ks[in.size-1]
     in.ks = in.ks[:in.size-1]
-    in.size = uint64(len(in.ks))
+    in.size = len(in.ks)
     // Return the median
     return median
 }
 
-func (child *inode) rotateLeft(in_ Node, count uint64, median bkey) bkey {
+func (child *inode) rotateLeft(in_ Node, count int, median bkey) bkey {
     in := in_.(*knode)
     child.ks = append(child.ks, median)
-    chlen := uint64(len(child.ks))
+    chlen := len(child.ks)
     // Move first `count` keys from in -> child.
     child.ks = child.ks[:chlen+count]  // First expand
     copy( child.ks[chlen:], in.ks[0:count] )
     // Blindly shrink in keys
     in.ks = in.ks[count:]
     // Update size.
-    in.size, child.size = uint64(len(in.ks)), uint64(len(child.ks))
+    in.size, child.size = len(in.ks), len(child.ks)
     // Move last count values from in -> child
     child.vs = child.vs[:chlen+count]    // First expand
     copy( child.vs[chlen:], in.vs[0:count] )
@@ -177,7 +177,7 @@ func (child *inode) rotateLeft(in_ Node, count uint64, median bkey) bkey {
     // Pop out median
     median = child.ks[child.size-1]
     child.ks = child.ks[:child.size-1]
-    child.size = uint64(len(child.ks))
+    child.size = len(child.ks)
     // Return the median
     return median
 }
